refactor(config): drain toml/mod.rs to module wiring only
Move the cross-cutting helpers out of mod.rs to the stage they belong to:
runtime extension traits and env/host/secret resolution to resolve.rs, path
sanitizing and `ValueOrUnlimited` to common.rs, and the data/cache dir
constants to server.rs. Each stage module now carries its own imports instead
of inheriting mod.rs's `use` block through `use super::*`.

mod.rs is now just the submodule declarations, re-exports, and `mod tests;`.
