+++ src/config/toml/authored.rs
@@ -4,6 +4,19 @@
 //! and verifies them.
 
 use super::*;
+use crate::args::TomlComponentType;
+use crate::config::env_var::EnvVarConfig;
+use anyhow::{Context, anyhow, bail};
+use concepts::cas::Cas;
+use concepts::component_id::ComponentDigest;
+use concepts::{ContentDigest, FunctionFqn};
+use hashbrown::HashMap;
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+use serde_with::{DeserializeFromStr, SerializeDisplay};
+use std::collections::BTreeMap;
+use std::fmt::Display;
+use std::str::FromStr;
 
 #[derive(Deserialize, Serialize, JsonSchema, Default, Clone)]
 #[serde(deny_unknown_fields)]
